// resources/js/api/generated/models/network.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { NetworkContainers } from './networkContainers';
import type { NetworkIPAM } from './networkIPAM';
import type { NetworkLabels } from './networkLabels';
import type { NetworkOptions } from './networkOptions';

export interface Network {
  containers?: NetworkContainers;
  created?: string;
  driver?: string;
  exists: boolean;
  external?: boolean;
  ipam?: NetworkIPAM | null;
  labels?: NetworkLabels;
  name: string;
  options?: NetworkOptions;
}