// resources/js/api/generated/models/stackPermissionsResponse.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { StackPermissionsData } from './stackPermissionsData';

export interface StackPermissionsResponse {
  data: StackPermissionsData;
  success: boolean;
}