// resources/js/api/generated/models/rawComposeConfig.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { RawComposeConfigConfigs } from './rawComposeConfigConfigs';
import type { RawComposeConfigNetworks } from './rawComposeConfigNetworks';
import type { RawComposeConfigSecrets } from './rawComposeConfigSecrets';
import type { RawComposeConfigServices } from './rawComposeConfigServices';
import type { RawComposeConfigVolumes } from './rawComposeConfigVolumes';

export interface RawComposeConfig {
  compose_file: string;
  configs?: RawComposeConfigConfigs;
  networks?: RawComposeConfigNetworks;
  secrets?: RawComposeConfigSecrets;
  services: RawComposeConfigServices;
  volumes?: RawComposeConfigVolumes;
}